refactor: Change parameter.
