+++ src/admin/pages/collections/Edit/index.tsx
@@ -1,10 +1,15 @@
 import React from 'react';
 import { useParams } from 'react-router-dom';
+import { Props } from './types';
 
-const EditPage: React.FC = () => {
+const EditPage: React.FC<Props> = ({ collection }) => {
   const { id } = useParams();
 
-  return <>{id ? 'Edit' : 'Create'} collection</>;
+  return (
+    <>
+      {id ? 'Edit' : 'Create'} {collection.collection}
+    </>
+  );
 };
 
 export default EditPage;
